requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
boto3
fastapi
uvicorn
python-multipart
pydantic
requests
mutagen
langchain
langchain-openai